requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
httpx
slowapi
beautifulsoup4
requests
supabase
passlib[bcrypt]
